requirements.txt
Scrapy>=1.4.0
scrapy-inline-requests>=0.3.1
